requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
selenium>=4.5.0
webdriver-manager>=4.0.0
